+++ src/core/shared/enums/itemType.ts
@@ -7,5 +7,6 @@ export enum ItemType {
     IS_TOOLBAR = 16,
     IS_WEAPON = 32,
     DESTROY_ON_DROP = 64,
-    CONSUMABLE = 128
+    CONSUMABLE = 128,
+    SKIP_CONSUMABLE = 256
 }