+++ src/core/shared/enums/system.ts
@@ -17,6 +17,7 @@ export enum SYSTEM_EVENTS {
     PLAYER_EMIT_SOUND_3D = 'sound:3D',
     PLAYER_EMIT_FRONTEND_SOUND = 'sound:Frontend',
     PLAYER_EMIT_NOTIFICATION = 'notification:Show',
+    PLAYER_EMIT_TASK_MOVE = 'task:Move',
     //
     PLAYER_RELOAD = 'player:ForceReload',
     //